+++ examples/map_basic.rs
@@ -25,7 +25,5 @@ fn startup(mut commands: Commands, asset_server: Res<AssetServer>) {
     let mut map_entity = commands.spawn(TiledMapHandle(map_handle));
 
     // You can eventually add some extra settings to your map
-    map_entity.insert(TiledMapSettings::with_layer_positioning(
-        LayerPositioning::Centered,
-    ));
+    map_entity.insert(TiledMapAnchor::Center);
 }