+++ examples/map_reload.rs
@@ -39,7 +39,7 @@ fn startup(
 
     commands.spawn((
         TiledMapHandle(asset_server.load("maps/orthogonal/finite.tmx")),
-        TiledMapSettings::with_layer_positioning(LayerPositioning::Centered),
+        TiledMapAnchor::Center,
     ));
     next_state.set(MapState::Loaded);
 }
@@ -61,7 +61,7 @@ fn handle_load(
         info!("Load map");
         commands.spawn((
             TiledMapHandle(asset_server.load("maps/orthogonal/finite.tmx")),
-            TiledMapSettings::with_layer_positioning(LayerPositioning::Centered),
+            TiledMapAnchor::Center,
         ));
         next_state.set(MapState::Loaded);
     }