+++ examples/orientation_hexagonal.rs
@@ -25,9 +25,7 @@ fn startup(mut commands: Commands, asset_server: Res<AssetServer>) {
     commands.spawn(Camera2d);
 
     let default_callback: helper::assets::MapInfosCallback = |c| {
-        c.insert(TiledMapSettings::with_layer_positioning(
-            LayerPositioning::Centered,
-        ));
+        c.insert(TiledMapAnchor::Center);
     };
 
     let mut mgr = helper::assets::AssetsManager::new(&mut commands);