+++ examples/physics_rapier_orientation.rs
@@ -31,7 +31,7 @@ fn startup(mut commands: Commands, asset_server: Res<AssetServer>) {
 
     let default_callback: helper::assets::MapInfosCallback = |c| {
         c.insert((
-            TiledMapSettings::with_layer_positioning(LayerPositioning::Centered),
+            TiledMapAnchor::Center,
             // For isometric maps, it can be useful to tweak bevy_ecs_tilemap render settings.
             // TilemapRenderSettings provide the 'y_sort' parameter to sort chunks using their y-axis
             // position during rendering.